Move summary page components in separate folder
