--- newdle/client/src/components/NewdleHeader.module.scss
@@ -1,23 +1,0 @@
-@import '../styles/palette';
-
-.title {
-  color: $purple;
-  display: flex;
-  align-items: center;
-
-  .header {
-    margin-bottom: 0;
-  }
-
-  .label {
-    margin-left: 5px;
-  }
-}
-
-.subtitle {
-  color: $white;
-}
-
-.summary {
-  padding: 30px 3em;
-}
