+++ newdle/client/src/components/summary/SummaryPage.js
@@ -1,14 +1,14 @@
 import React, {useState} from 'react';
-import {serializeDate, toMoment} from '../util/date';
+import {serializeDate, toMoment} from '../../util/date';
 import {useDispatch, useSelector} from 'react-redux';
 import {Button, Container, Icon, Loader} from 'semantic-ui-react';
-import ParticipantTable from './ParticipantTable';
-import {getNewdle} from '../selectors';
-import {updateNewdle} from '../actions';
-import client from '../client';
-import styles from './Summary.module.scss';
+import ParticipantTable from '../ParticipantTable';
+import {getNewdle} from '../../selectors';
+import {updateNewdle} from '../../actions';
+import client from '../../client';
+import styles from './summary.module.scss';
 
-export default function Summary() {
+export default function SummaryPage() {
   const [submitting, setSubmitting] = useState(false);
   const [error, setError] = useState('');
   const [finalDate, setFinalDate] = useState(null);
@@ -31,12 +31,30 @@ export default function Summary() {
   };
 
   if (!newdle) {
-    return <Loader />;
+    return <Loader active />;
   }
 
   return (
     <Container text>
-      {!newdle.final_dt && (
+      {newdle.final_dt ? (
+        <>
+          <div className={styles['summary-container']}>
+            <div className={styles.datetime}>
+              <Icon name="calendar alternate outline" />
+              {serializeDate(newdle.final_dt, 'MMMM Do YYYY')}
+            </div>
+            <div className={styles.datetime}>
+              <Icon name="clock outline" />
+              {serializeDate(newdle.final_dt, 'h:mm')} -{' '}
+              {serializeDate(toMoment(newdle.final_dt).add(newdle.duration, 'm'), 'h:mm')} (
+              {newdle.timezone})
+            </div>
+          </div>
+          <div className={styles['button-row']}>
+            <Button className={styles['create-event-button']}>Create event</Button>
+          </div>
+        </>
+      ) : (
         <>
           <ParticipantTable finalDate={finalDate} setFinalDate={setFinalDate} />
           <div className={styles['button-row']}>
@@ -49,25 +67,6 @@ export default function Summary() {
             >
               Select final date
             </Button>
-          </div>
-        </>
-      )}
-      {newdle.final_dt && (
-        <>
-          <div className={styles['summary-container']}>
-            <div className={styles['datetime']}>
-              <Icon name="calendar alternate outline" />
-              {serializeDate(newdle.final_dt, 'MMMM Do YYYY')}
-            </div>
-            <div className={styles['datetime']}>
-              <Icon name="clock outline" />
-              {serializeDate(newdle.final_dt, 'h:mm')} -{' '}
-              {serializeDate(toMoment(newdle.final_dt).add(newdle.duration, 'm'), 'h:mm')} (
-              {newdle.timezone})
-            </div>
-          </div>
-          <div className={styles['button-row']}>
-            <Button className={styles['create-event-button']}>Create event</Button>
           </div>
         </>
       )}
