+++ newdle/client/src/components/summary/index.js
@@ -1,0 +1,2 @@
+export {default as SummaryHeader} from './SummaryHeader';
+export {default as SummaryPage} from './SummaryPage';
